requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.3
Pillow==10.1.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
